chore(dts): split usage analyzer symbol tracking
